+++ sobek-common/src/main/java/org/rutebanken/sobek/netex/util/PublicationDeliveryUnmarshaller.java
@@ -13,7 +13,7 @@
  * limitations under the Licence.
  */
 
-package org.rutebanken.sobek.rest.netex.publicationdelivery;
+package org.rutebanken.sobek.netex.util;
 
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBElement;
