+++ sobek-common/src/test/java/org/rutebanken/sobek/SobekTestApplication.java
@@ -1,0 +1,11 @@
+package org.rutebanken.sobek;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SobekTestApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(SobekTestApplication.class, args);
+  }
+}
